refactor: rename to just secrets
